menu/characterとgiftもpinia更新

+++ src/server/usePlayerID.ts
@@ -1,7 +1,9 @@
+import { onMounted } from "vue";
 import { db } from "./firebase";
 import { collection, doc, addDoc, getDoc, getDocs, deleteDoc } from "firebase/firestore";
 import type { PlayerData, Character, Gift } from "@/types";
 import { playerStore } from "@/main";
+import { storeToRefs } from "pinia";
 
 //Collectionの参照
 const playersRef = collection(db, "players");
@@ -9,6 +11,10 @@ const charactersRef = collection(db, "characters");
 const giftsRef = collection(db, "gifts");
 
 //piniaの参照
+onMounted(async () => {
+  //!どうにかしてplayerStoreのidを取得したい
+  const { id } = storeToRefs(playerStore);
+});
 
 //player登録
 async function registerPlayer(): Promise<void> {
@@ -23,6 +29,7 @@ async function registerPlayer(): Promise<void> {
 //player削除
 async function deletePlayer(): Promise<void> {
   try {
+    if(!playerStore.id) return;
     await deleteDoc(doc(playersRef, playerStore.id));
     console.log("Player deleted: ", playerStore.id);
   } catch (error) {
